refactor: :closure signature to :value + :stop
Switch the `:closure` defsys signature to return a map with `:value` and
`:stop` that can be used to allocate and release resources.

Upon attempting to use the old version, I found myself needing atoms for
references that would be allocated in start and cleaned up in stop. Now
`:closure` just functions as start directly and we can relying on
capturing the variables for a stop function.